requirements.txt
flask
flask-sqlalchemy
google-genai
python-dotenv
flask-login
Pillow
